size prototype entropy budget from arbitrary size hint
