+++ soroban-sdk-macros/src/arbitrary.rs
@@ -317,11 +317,11 @@ fn quote_arbitrary(
     // prototype be named directly in the parameter list of a proptest! test.
     let proptest_tokens = if cfg!(feature = "testutils-proptest") {
         quote! {
-            impl #path::testutils::arbitrary::proptest::arbitrary::Arbitrary for #arbitrary_type_ident {
+            impl #path::testutils::proptest::proptest::arbitrary::Arbitrary for #arbitrary_type_ident {
                 type Parameters = ();
-                type Strategy = #path::testutils::arbitrary::proptest_arbitrary_interop::ArbStrategy<Self>;
+                type Strategy = #path::testutils::proptest::proptest_arbitrary_interop::ArbStrategy<Self>;
                 fn arbitrary_with(_: Self::Parameters) -> Self::Strategy {
-                    #path::testutils::arbitrary::proptest_arbitrary_interop::arb::<Self>()
+                    #path::testutils::proptest::arb_from_size_hint::<Self>()
                 }
             }
         }
