+++ soroban-sdk/src/testutils/proptest.rs
@@ -93,12 +93,22 @@
 //!
 //! ## Entropy budget
 //!
-//! `arb` feeds a fixed 256 bytes of random data to the `Arbitrary`
-//! implementation of the prototype. When a prototype needs more bytes than
-//! that, `arbitrary` zero-fills the remainder rather than returning an error,
-//! so the tail of a large prototype — for example the last fields of a
-//! `contracttype` struct with many `BytesN<32>` fields — is all zeros in every
-//! generated value. Use `arb_sized` with a larger budget for such types.
+//! Values are generated by feeding random bytes to the `Arbitrary`
+//! implementation of the prototype, and when the bytes run out `arbitrary`
+//! zero-fills the rest rather than returning an error, so too small a budget
+//! produces values with a constant tail.
+//!
+//! The `Arbitrary` implementations take the budget from the prototype's
+//! `arbitrary` size hint, which is exact for every prototype whose size is
+//! bounded, including the ones `contracttype` generates. A `contracttype`
+//! struct with twelve `BytesN<32>` fields is given the 384 bytes it needs.
+//! Collections, strings, and `Val` have no bounded size — more bytes means
+//! longer collections rather than a more complete value — so they are given a
+//! 256 byte default.
+//!
+//! `arb` uses a fixed 256 bytes regardless of the type, so a large prototype
+//! generated with it has a zeroed tail. Use `arb_sized` to raise the budget, or
+//! name the prototype and let the size hint do it.
 //!
 //! ## Limits of generated values
 //!
@@ -114,6 +124,10 @@ use core::fmt::Debug;
 
 /// A reexport of the `proptest` crate.
 pub use ::proptest;
+
+// Used by `contracttype` to implement proptest's `Arbitrary` on the prototype it generates.
+#[doc(hidden)]
+pub use ::proptest_arbitrary_interop;
 
 /// A [`proptest`] strategy that generates prototypes of the Soroban type `T`.
 ///
@@ -212,4 +226,145 @@ where
     T::Prototype: Debug + Clone + 'static,
 {
     proptest_arbitrary_interop::arb_sized::<T::Prototype>(size)
+}
+
+/// Implementations of [`proptest`]'s `Arbitrary` trait for the prototypes.
+///
+/// The prototypes already implement [`arbitrary`]'s `Arbitrary`, and these
+/// implementations forward to it, so that a prototype can be named directly in
+/// the parameter list of a `proptest!` test, in the same way that a fuzz test
+/// names it in its input struct.
+///
+/// The prototypes of the scalar types are the scalar types themselves, and
+/// `proptest` already implements the trait for them. The prototype of a tuple
+/// is a tuple of prototypes, which `proptest` covers with its own tuple
+/// implementations.
+///
+/// The entropy budget given to a prototype whose size the `arbitrary`
+/// implementation does not bound, such as a collection or a `Val`. There is no
+/// size that is correct for these, because more bytes means longer collections
+/// rather than a more complete value, so this is the same default that
+/// `proptest-arbitrary-interop` uses.
+#[doc(hidden)]
+pub const UNBOUNDED_ENTROPY_BUDGET: usize = proptest_arbitrary_interop::DEFAULT_SIZE;
+
+/// A strategy that generates the prototype `A` with an entropy budget taken
+/// from its `arbitrary` implementation's size hint.
+///
+/// Values are generated by feeding random bytes to the `arbitrary`
+/// implementation, and when the bytes run out `arbitrary` zero-fills the rest
+/// instead of erroring, so a budget smaller than the prototype needs produces a
+/// value with a constant tail. The size hint gives the exact number of bytes
+/// for prototypes with a bounded size, which is every prototype except the
+/// collections, the strings, and `Val`.
+///
+/// Used by the `Arbitrary` implementations below and by the ones that
+/// `contracttype` generates.
+#[doc(hidden)]
+pub fn arb_from_size_hint<A>() -> proptest_arbitrary_interop::ArbStrategy<A>
+where
+    A: proptest_arbitrary_interop::ArbInterop,
+{
+    let (lower, upper) = <A as ::arbitrary::Arbitrary>::size_hint(0);
+    let size = upper.unwrap_or(UNBOUNDED_ENTROPY_BUDGET).max(lower);
+    proptest_arbitrary_interop::arb_sized::<A>(size)
+}
+
+/// [`proptest`]: https://github.com/proptest-rs/proptest/
+mod prototype_impls {
+    use proptest_arbitrary_interop::{ArbInterop, ArbStrategy};
+
+    use super::arb_from_size_hint;
+    use crate::testutils::arbitrary::composite::*;
+    use crate::testutils::arbitrary::objects::*;
+
+    /// Implement `proptest`'s `Arbitrary` for prototypes that have no type
+    /// parameters.
+    macro_rules! impl_arbitrary {
+        ($($t:ty,)*) => {
+            $(
+                impl proptest::arbitrary::Arbitrary for $t {
+                    type Parameters = ();
+                    type Strategy = ArbStrategy<Self>;
+                    fn arbitrary_with(_: Self::Parameters) -> Self::Strategy {
+                        arb_from_size_hint::<Self>()
+                    }
+                }
+            )*
+        };
+    }
+
+    impl_arbitrary! {
+        ArbitraryU256,
+        ArbitraryI256,
+        ArbitraryBytes,
+        ArbitraryString,
+        ArbitrarySymbol,
+        ArbitraryAddress,
+        ArbitraryMuxedAddress,
+        ArbitraryTimepoint,
+        ArbitraryDuration,
+        ArbitraryBls12381Fp,
+        ArbitraryBls12381Fp2,
+        ArbitraryBls12381G1Affine,
+        ArbitraryBls12381G2Affine,
+        ArbitraryBls12381Fr,
+        ArbitraryBn254G1Affine,
+        ArbitraryBn254G2Affine,
+        ArbitraryBn254Fp,
+        ArbitraryBn254Fr,
+        ArbitraryVal,
+        ArbitraryValVec,
+        ArbitraryValMap,
+        ArbitraryValOption,
+    }
+
+    // The prototypes that have type parameters are implemented individually.
+    // The `Self: ArbInterop` bound defers the requirements on the type
+    // parameters to the prototype's `arbitrary` implementation, so that they do
+    // not need restating here.
+
+    impl<T> proptest::arbitrary::Arbitrary for ArbitraryOption<T>
+    where
+        Self: ArbInterop,
+    {
+        type Parameters = ();
+        type Strategy = ArbStrategy<Self>;
+        fn arbitrary_with(_: Self::Parameters) -> Self::Strategy {
+            arb_from_size_hint::<Self>()
+        }
+    }
+
+    impl<T> proptest::arbitrary::Arbitrary for ArbitraryVec<T>
+    where
+        Self: ArbInterop,
+    {
+        type Parameters = ();
+        type Strategy = ArbStrategy<Self>;
+        fn arbitrary_with(_: Self::Parameters) -> Self::Strategy {
+            arb_from_size_hint::<Self>()
+        }
+    }
+
+    impl<K, V> proptest::arbitrary::Arbitrary for ArbitraryMap<K, V>
+    where
+        Self: ArbInterop,
+    {
+        type Parameters = ();
+        type Strategy = ArbStrategy<Self>;
+        fn arbitrary_with(_: Self::Parameters) -> Self::Strategy {
+            arb_from_size_hint::<Self>()
+        }
+    }
+
+    impl<const N: usize> proptest::arbitrary::Arbitrary for ArbitraryBytesN<N>
+    where
+        Self: ArbInterop,
+    {
+        type Parameters = ();
+        type Strategy = ArbStrategy<Self>;
+        fn arbitrary_with(_: Self::Parameters) -> Self::Strategy {
+            arb_from_size_hint::<Self>()
+        }
+    }
 }
